requirements.txt
fastapi
uvicorn
numpy
librosa
resampy
scikit-learn
joblib
pydantic
python-multipart
python-dotenv
pydantic-settings
elevenlabs